+++ Assets/Scripts/MoveInfo.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class MoveInfo : MonoBehaviour
 {
     private string moveName;
-    private int damage;
+    private float damage;
     private float speed;
     private int knockback;
     private MoveType type;
     private RangeType rtype;
     private GameObject visual;
-    public MoveInfo (string moveName, int damage, float speed, int knockback, MoveType type, RangeType rtype, GameObject visual)
+    public MoveInfo (string moveName, float damage, float speed, int knockback, MoveType type, RangeType rtype, GameObject visual)
     {
         this.moveName = moveName;
         this.damage = damage;
@@ -26,7 +26,7 @@ public class MoveInfo : MonoBehaviour
         return this.moveName;
     }
 
-    public int GetDamage()
+    public float GetDamage()
     {
         return this.damage;
     }
